+++ src/hooks/index.ts
@@ -1,8 +1,10 @@
 export * from './useAutoScroll';
 export * from './useCueAlignment';
 export * from './useCueEditor';
+export * from './useEscapeKey';
 export * from './useKeyboardShortcuts';
 export * from './useScriptPreferences';
 export * from './useScriptStorage';
 export * from './useScriptTheme';
 export * from './useYouTubePlayer';
+